backend/x64: Allow fsizes of 16 in ChooseOnFsize
Given half-precision instructions aren't present on x64, we can allow
the size, but assert on it, given all half-precision instructions are
going to have to fall back to softfloat implementations.

This allows functions that have a fast path (that uses ChooseOnFsize),
and a softfloat fallback to be easily usable with half-precision
floating point overloads.